Random Trees 1
This is a start to what I could translate from the mathematica code. I just can't figure out how to reorganize the numbers. But for the most part this makes random trees with bifurcations or trifurcations. I am also trying to figure out how to make the the number of initial connections also random. Andddd I am trying to figure out the if else for while loops that can get rid of nodes and branches as the fire spreads.